Moved new version QA notebooks into old file names
