+++ qa/templates/config.py
@@ -1,21 +1,10 @@
-DATASET_OLD = 'world-fishing-827.pipe_production_v20201001'
-DATASET_NEW_MONTHLY_INT = 'pipe_ais_test_20220901_monthly_internal'
-DATASET_NEW_BACKFILL_INT = 'pipe_ais_test_20220901_backfill_internal'
-DATASET_NEW_MONTHLY_PUB = 'pipe_ais_test_20220901_monthly_published'
-DATASET_NEW_BACKFILL_PUB = 'pipe_ais_test_20220901_backfill_published'
+PROJECT = 'world-fishing-827'
+DATASET_PIPE3 = 'pipe_ais_v3_alpha_internal'
+DATASET_PIPE25 = 'pipe_production_v20201001'
 
-SAT_OFFSETS_TABLE = 'sat_time_offsets_'
-MESSAGES_SEGMENTED_TABLE = 'messages_segmented_'
-FRAGMENTS_TABLE = 'fragments_'
 SEGMENTS_TABLE = 'segments_'
-SEGMENT_IDENTITY_DAILY_TABLE = 'segment_identity_daily_'
-SEGMENT_VESSEL_DAILY_TABLE = 'segment_vessel_daily_'
+FRAGMENTS_TABLE = 'fragments_'
+MESSAGES_SEGMENTED_TABLE = 'messages_segmented_'
 SEGMENT_INFO_TABLE = 'segment_info'
-SEGMENT_VESSEL_TABLE = 'segment_vessel'
-VESSEL_INFO_TABLE = 'vessel_info'
+SEGMENT_IDENITY_DAILY_TABLE = 'segment_identity_daily_'
 
-# Threshold used for the difference between the
-# number of segments in a given day compared to
-# the rolling 30-day average.
-# Used in `segmenter_version_comparison.py`.
-NUM_SEGS_DIFF_THRESHOLD = 0.2
