--- qa/templates/config_new.py
@@ -1,10 +1,0 @@
-PROJECT = 'world-fishing-827'
-DATASET_PIPE3 = 'pipe_ais_v3_alpha_internal'
-DATASET_PIPE25 = 'pipe_production_v20201001'
-
-SEGMENTS_TABLE = 'segments_'
-FRAGMENTS_TABLE = 'fragments_'
-MESSAGES_SEGMENTED_TABLE = 'messages_segmented_'
-SEGMENT_INFO_TABLE = 'segment_info'
-SEGMENT_IDENITY_DAILY_TABLE = 'segment_identity_daily_'
-
